Python/grpc_server/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.25.1
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2